Fix so that www.prnewswire.com works
The website http://www.prnewswire.com/ has some articles which have uneven
anchor elements (they are missing the starting <a>). I added a
check for this in TagActionAnchor.End() and I removed the panic() during
checking in TagActionAnchor.Start(). The result of this bug was that
LinkDensity would be incorrect further down the line causing
mis-classification of TextBlocks.

The article
http://www.prnewswire.com/news-releases/bascom-closes-36-unit-apartment-community-in-expanding-inland-empire-300390850.html
is an example of an article in which content was not being extracted
properly.